+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2024-05-10
+## [Unreleased] - 2024-05-13
 
 **BREAKING CHANGES: See Removed and Changed sections for details.**
 
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   * Streams of random triples of distinct sorted integers (method sortedTriples)
   * Streams of random triples of distinct integers, separated by at most a specified window (method windowedTriples)
   * Streams of random triples of distinct sorted integers, separated by at most a specified window (method sortedWindowedTriples)
+* Shuffler class for efficiently randomizing the ordering of elements in arrays.
 
 ### Changed
 * Refactored and optimized RandomIndexer.nextIntTriple methods.